+++ includes/pis-core.php
@@ -26,9 +26,10 @@ if ( ! defined( 'WPINC' ) ) {
 /**
  * The core function.
  *
- * @since 1.0
  * @param array $args The options for the main function.
  * @return string The HTML output.
+ * @since 1.0
+ * @since 4.15.0 Added compatibility with Yoast SEO plugin.
  */
 function pis_get_posts_in_sidebar( $args ) {
 	$args = wp_parse_args( $args, pis_get_defaults() );
@@ -355,16 +356,24 @@ function pis_get_posts_in_sidebar( $args ) {
 	 */
 
 	/*
-	 * Check if the user wants to display posts from the same category of the single post.
-	 * The parameters for excluding posts (like "post__not_in") will be left active.
+	 * Check if the user wants to display posts from the same category of the
+	 * single post.
+	 * The parameters for excluding posts (like "post__not_in") will be left
+	 * active.
 	 * This will work in single (regular) posts and in custom post types.
 	 *
-	 * The category used will be the first in the array ($post_categories[0]), i.e. the category with the lowest ID.
-	 * In the permalink WordPress uses the category with the lowest ID and we want to use this.
-	 * On the contrary, if we get the list of posts categories, these are returned by WordPress in an aphabetically ordered array,
-	 * where the lowest key ID has not always the category used in the permalink.
+	 * The category used will be the first in the array ($post_categories[0]),
+	 * i.e. the category with the lowest ID.
+	 * In the permalink, WordPress uses the category with the lowest ID and we
+	 * want to use this.
+	 * On the contrary, if we get the list of posts categories, these are
+	 * returned by WordPress in an aphabetically ordered array,
+	 * where the lowest key ID has not always the category used in the
+	 * permalink.
 	 *
 	 * @since 3.2
+	 * @since 4.15.0 Added compatibility with Yoast SEO plugin, which lets the
+	 *               user to choose a main category.
 	 */
 	if ( $get_from_same_cat && is_single() ) {
 		// Set the post_type.
@@ -380,9 +389,18 @@ function pis_get_posts_in_sidebar( $args ) {
 		}
 
 		// Set the category.
-		$post_categories = wp_get_post_categories( $single_post_id );
+		if ( $yoast_main_cat && function_exists( 'yoast_get_primary_term_id' ) && false !== yoast_get_primary_term_id( 'category', $single_post_id ) ) {
+			$post_categories = explode( ',', yoast_get_primary_term_id( 'category', $single_post_id ) );
+		} else {
+			$post_categories = wp_get_post_categories( $single_post_id );
+		}
 		if ( $post_categories ) {
-			// Sort the categories of the post in ascending order, so to use the category used by WordPress in the permalink.
+			/* Sort the categories of the post in ascending order based on the
+			 * ID of the categories, so to use the category used by WordPress in
+			 * the permalink.
+			 * $sort_categories has no effect if $yoast_main_cat is active,
+			 * because $yoast_main_cat is always an array of one element.
+			 */
 			if ( $sort_categories ) {
 				sort( $post_categories );
 			}
@@ -1057,7 +1075,7 @@ function pis_get_posts_in_sidebar( $args ) {
 	if ( $pis_query->have_posts() ) :
 
 		if ( $intro ) {
-			$pis_output = '<p ' . pis_paragraph( $intro_margin, $margin_unit, 'pis-intro', 'pis_intro_class' ) . '>' . pis_break_text( $intro ) . '</p>';
+			$pis_output .= "\n" . '<p ' . pis_paragraph( $intro_margin, $margin_unit, 'pis-intro', 'pis_intro_class' ) . '>' . pis_break_text( $intro ) . '</p>';
 		}
 
 		// When updating from 1.14, the $list_element variable is empty.
@@ -1095,7 +1113,7 @@ function pis_get_posts_in_sidebar( $args ) {
 		 * Add the ID selector to UL since some page builder plugins remove the section HTML tag.
 		 * @since 4.5.0
 		 */
-		$pis_output .= '<' . $list_element . $pis_ul_id . pis_class( 'pis-ul', apply_filters( 'pis_ul_class', '' ), false ) . $bullets_style . '>' . "\n";
+		$pis_output .= "\n" . '<' . $list_element . $pis_ul_id . pis_class( 'pis-ul', apply_filters( 'pis_ul_class', '' ), false ) . $bullets_style . '>' . "\n";
 
 		while ( $pis_query->have_posts() ) :
 			$pis_query->the_post();
@@ -1152,7 +1170,7 @@ function pis_get_posts_in_sidebar( $args ) {
 					$wp_post_classes = ' ' . implode( ' ', get_post_class( '', $pis_query->post->ID ) );
 				}
 
-				$pis_output .= '<li ' . pis_class( 'pis-li' . $post_id_class . $current_post_class . $sticky_class . $private_class . $wp_post_classes, apply_filters( 'pis_li_class', '' ), false ) . '>' . "\n";
+				$pis_output .= "\t" . '<li ' . pis_class( 'pis-li' . $post_id_class . $current_post_class . $sticky_class . $private_class . $wp_post_classes, apply_filters( 'pis_li_class', '' ), false ) . '>' . "\n";
 
 				// Define the containers for single sections to be concatenated later.
 				$pis_thumbnail_content    = '';
@@ -1180,7 +1198,7 @@ function pis_get_posts_in_sidebar( $args ) {
 						'image_link'          => $image_link,
 						'image_link_to_post'  => $image_link_to_post,
 					)
-				) . "\n";
+				);
 
 				/* The title */
 				$pis_title_content .= pis_the_title(
@@ -1199,7 +1217,7 @@ function pis_get_posts_in_sidebar( $args ) {
 						'title_hellipsis'    => $title_hellipsis,
 						'html_title_type_of' => $html_title_type_of,
 					)
-				) . "\n";
+				);
 
 				/*
 				 * The post content.
@@ -1224,7 +1242,7 @@ function pis_get_posts_in_sidebar( $args ) {
 
 					// If the the text of the post is empty or the user does not want to display the image, hide the HTML p tag.
 					if ( ! empty( $pis_the_text ) || ( $display_image && ! $image_before_title ) ) {
-						$pis_text_content .= '<p ' . pis_paragraph( $excerpt_margin, $margin_unit, 'pis-excerpt', 'pis_excerpt_class' ) . '>';
+						$pis_text_content .= "\n\t\t" . '<p ' . pis_paragraph( $excerpt_margin, $margin_unit, 'pis-excerpt', 'pis_excerpt_class' ) . '>';
 					}
 
 					if ( $display_image && ! $image_before_title ) {
@@ -1266,7 +1284,7 @@ function pis_get_posts_in_sidebar( $args ) {
 					$pis_text_content .= $pis_the_text;
 
 					if ( ! empty( $pis_the_text ) || ( $display_image && ! $image_before_title ) ) {
-						$pis_text_content .= '</p>' . "\n";
+						$pis_text_content .= '</p>';
 					}
 				endif;
 				/* Close the post content */
@@ -1304,7 +1322,7 @@ function pis_get_posts_in_sidebar( $args ) {
 						'gravatar_size'         => $gravatar_size,
 						'gravatar_default'      => $gravatar_default,
 					)
-				) . "\n";
+				);
 
 				/* The categories */
 				$pis_categories_content .= pis_the_categories(
@@ -1315,7 +1333,7 @@ function pis_get_posts_in_sidebar( $args ) {
 						'margin_unit'       => $margin_unit,
 						'categ_text'        => $categ_text,
 					)
-				) . "\n";
+				);
 
 				/* The tags */
 				$pis_tags_content .= pis_the_tags(
@@ -1327,7 +1345,7 @@ function pis_get_posts_in_sidebar( $args ) {
 						'margin_unit' => $margin_unit,
 						'tags_text'   => $tags_text,
 					)
-				) . "\n";
+				);
 
 				/* The custom taxonomies */
 				$pis_custom_tax_content .= pis_custom_taxonomies_terms_links(
@@ -1338,7 +1356,7 @@ function pis_get_posts_in_sidebar( $args ) {
 						'terms_margin' => $terms_margin,
 						'margin_unit'  => $margin_unit,
 					)
-				) . "\n";
+				);
 
 				/* The post meta */
 				$pis_custom_field_content .= pis_custom_field(
@@ -1354,7 +1372,7 @@ function pis_get_posts_in_sidebar( $args ) {
 						'custom_field_margin' => $custom_field_margin,
 						'margin_unit'         => $margin_unit,
 					)
-				) . "\n";
+				);
 
 				// Concatenate the variables.
 				if ( $display_image && $image_before_title ) {
@@ -1416,7 +1434,7 @@ function pis_get_posts_in_sidebar( $args ) {
 					$pis_output .= $pis_custom_field_content;
 				}
 
-				$pis_output .= '</li>' . "\n";
+				$pis_output .= "\n\t" . '</li>' . "\n";
 				/* Close li */
 
 			}
@@ -1440,6 +1458,7 @@ function pis_get_posts_in_sidebar( $args ) {
 					'margin_unit'     => $margin_unit,
 					'post_id'         => $single_post_id,
 					'sort_categories' => $sort_categories,
+					'yoast_main_cat'  => $yoast_main_cat,
 					'sort_tags'       => $sort_tags,
 				)
 			);